Testing and fixing a load of rubbish

+++ src/main/java/com/example/_40krusadebackend/Service/UnitDetailsService.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 
 public interface UnitDetailsService {
     Optional<UnitDetails> createUnit(UnitDetails unitDetails);
-    Optional<UnitDetails> getUnitById(Long unitDetailsId);
+    Optional<UnitDetails> getUnitById(int unitDetailsId);
     Optional<UnitDetails> findByUnitOfficialName(String unitOfficialName);
     List<UnitDetails> getAllUnits();
     UnitDetails updateUnit(int unitDetailsId, UnitDetails unitOfficialName);
-    void deleteUnit(Long unitDetailsId);
+    void deleteUnit(int unitDetailsId);
 }
